+++ src/devices/src/virtio/vsock/tcp.rs
@@ -229,7 +229,7 @@ impl TcpProxy {
     }
 
     fn peer_avail_credit(&self) -> usize {
-        (Wrapping(self.peer_buf_alloc as u32) - (self.rx_cnt - self.peer_fwd_cnt)).0 as usize
+        (Wrapping(self.peer_buf_alloc) - (self.rx_cnt - self.peer_fwd_cnt)).0 as usize
     }
 
     fn recv_to_pkt(&self, pkt: &mut VsockPacket) -> RecvPkt {
